refactor polyglot a bit more
